test: nothingness -> @nothingness/level -> nothingness
Symbols are a useful tool for encapsulation, but they suffer from the
same problems as anything else trying to enforce a singleton pattern.
I've modified the LevelDB adaptor to allow you to pass in the symbol it
should use for setting the ID on a loaded object, which is something
that the default serialization / deserialization helpers enforce (and
which overriding helpers should also enforce).

There's a pretty strict protocol to using this set of decoupled patterns
together. It's going to need some careful documentation to be usable.